fixes #2010 :: update intrinsics / GDC runtime support; lots of thanks to ibuclaw

+++ tango/net/device/Berkeley.d
@@ -2220,7 +2220,7 @@ public class SocketSet
                         if (s > _maxfd)
                                 _maxfd = s;
 
-                        bts(cast(uint*)&first[fdelt(s)], cast(uint)s % nfdbits);
+                        bts(cast(size_t*)&first[fdelt(s)], cast(size_t)s % nfdbits);
                 }
                 else
                 {
@@ -2258,7 +2258,7 @@ public class SocketSet
                 }
                 else version (Posix)
                 {
-                        btr(cast(uint*)&first[fdelt(s)], cast(uint)s % nfdbits);
+                        btr(cast(size_t*)&first[fdelt(s)], cast(size_t)s % nfdbits);
 
                         // If we're removing the biggest file descriptor we've
                         // entered so far we need to recalculate this value
